utilities/coverage/requirements.txt
python-gitlab==3.15.0
